@@ -3,5 +3,5 @@ import 'package:flutter/foundation.dart';
 @immutable
 class Constants {
   // ignore: unused_field
-  static Constants _instance;
+  static Constants? _instance;
 }
