@@ -4,13 +4,13 @@ import 'package:rcmasbusapp/data/remote/auth_data_source.dart';
 import 'package:rcmasbusapp/data/repository/auth_repository.dart';
 
 class AuthRepositoryImpl implements AuthRepository {
-  AuthRepositoryImpl({@required AuthDataSource authDataSource})
+  AuthRepositoryImpl({required AuthDataSource authDataSource})
       : _dataSource = authDataSource;
 
   final AuthDataSource _dataSource;
 
   @override
-  User fetchUser() {
+  User? fetchUser() {
     return _dataSource.getUser();
   }
 
@@ -25,7 +25,7 @@ class AuthRepositoryImpl implements AuthRepository {
   }
 
   @override
-  Future<void> submitOTP(String smsCode, String verificationId) {
+  Future<void> submitOTP(String smsCode, String? verificationId) {
     return _dataSource.submitOTP(smsCode, verificationId);
   }
 }
